+++ NEWS.md
@@ -7,7 +7,7 @@ affiner 2.0.0 (development)
   The `is_equivalent()` method for angle vectors tests whether two angles are congruent.
   The `is_equivalent()` method for `Point1D`, `Line2D`, `Plane3D` classes tests whether they are the same point/line/plane after standardization.
 * `is_parallel()` is a generic S3 method which tests whether two objects are "parallel" (e.g. `Line2D` and `Plane3D` objects).
-* `isotoxal_2ngon_inner_radius()` computes the inner radius of an isotoxal `2n`-gon polygon (#66).
+* `isotoxal_2ngon_inner_radius()` computes the inner radius of an isotoxal `2n`-gon polygon (#66).  `star_inner_radius()` is an alias.
 * If R >= 4.4.0 can now also use `crossprod()` (in addition to `cross_product3d()`) to compute the
   **vector** cross product of two `Coord3D` objects.
 * Can now use `ceiling()`, `floor()`, `round()`, `signif()`, and `trunc()` on `Coord1D`, `Coord2D`, and `Coord3D` objects (#63).